database/pyproject.toml
[tool.poetry]
name = "database"
version = "0.1.0"
description = ""
authors = ["Iury Rosal <iuryrosal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
faker = "^27.0.0"
pg8000 = "^1.31.2"
python-dotenv = "^1.0.1"
cloud-sql-python-connector = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
